requirements.txt
psycopg2-binary>=2.9.0
faker>=18.0.0
pandas>=1.5.0
openpyxl>=3.0.0
tabulate>=0.8.9